Crate bholdus_chainbridge

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

ProposalVotes

Enums

ProposalStatus

Functions

derive_resource_id
Helper function to concatenate a chain ID and some bytes to produce a resource ID. The common format is (31 bytes unique ID + 1 byte chain ID).

Type Definitions

ChainId

DepositNonce

ResourceId