Struct bholdus_chainbridge::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn set_threshold(origin: OriginFor<T>, threshold: u32) -> DispatchResult

Sets the vote threshold for proposals.

This threshold is used to determine how many votes are required before a proposal is executed.

O(1) lookup and insert

pub fn set_resource(
origin: OriginFor<T>,
id: ResourceId,
method: Vec<u8>
) -> DispatchResult

Stores a method name on chain under an associated resource ID.

O(1) write

pub fn remove_resource(origin: OriginFor<T>, id: ResourceId) -> DispatchResult

Removes a resource ID from the resource mapping.

After this call, bridge transfers with the associated resource ID will be rejected.

O(1) removal

pub fn whitelist_chain(origin: OriginFor<T>, id: ChainId) -> DispatchResult

Enables a chain ID as a source or destination for a bridge transfer.

O(1) lookup and insert

pub fn add_relayer(origin: OriginFor<T>, v: T::AccountId) -> DispatchResult

Adds a new relayer to the relayer set.

O(1) lookup and insert

pub fn remove_relayer(origin: OriginFor<T>, v: T::AccountId) -> DispatchResult

Removes an existing relayer from the set.

O(1) lookup and removal

pub fn acknowledge_proposal(
origin: OriginFor<T>,
nonce: DepositNonce,
src_id: ChainId,
r_id: ResourceId,
call: Box<<T as Config>::Proposal>
) -> DispatchResult

Commits a vote in favour of the provided proposal.

If a proposal with the given nonce and source chain ID does not already exist, it will be created with an initial vote in favour from the caller.

weight of proposed call, regardless of whether execution is performed

pub fn reject_proposal(
origin: OriginFor<T>,
nonce: DepositNonce,
src_id: ChainId,
r_id: ResourceId,
call: Box<<T as Config>::Proposal>
) -> DispatchResult

Commits a vote against a provided proposal.

Fixed, since execution of proposal should not be included

pub fn eval_vote_state(
origin: OriginFor<T>,
nonce: DepositNonce,
src_id: ChainId,
prop: Box<<T as Config>::Proposal>
) -> DispatchResult

Evaluate the state of a proposal given the current vote threshold.

A proposal with enough votes will be either executed or cancelled, and the status will be updated accordingly.

weight of proposed call, regardless of whether execution is performed

impl<T: Config> Pallet<T>

pub fn ensure_admin(o: T::Origin) -> DispatchResult

pub fn is_relayer(who: &T::AccountId) -> bool

Checks if who is a relayer

pub fn account_id() -> T::AccountId

Provides an AccountId for the pallet. This is used both as an origin check and deposit/withdrawal account.

pub fn resource_exists(id: ResourceId) -> bool

Asserts if a resource is registered

pub fn chain_whitelisted(id: ChainId) -> bool

Checks if a chain exists as a whitelisted destination

pub fn set_relayer_threshold(threshold: u32) -> DispatchResult

Set a new voting threshold

pub fn register_resource(id: ResourceId, method: Vec<u8>) -> DispatchResult

Register a method for a resource Id, enabling associated transfers

pub fn unregister_resource(id: ResourceId) -> DispatchResult

Removes a resource ID, disabling associated transfer

pub fn whitelist(id: ChainId) -> DispatchResult

Whitelist a chain ID for transfer

pub fn register_relayer(relayer: T::AccountId) -> DispatchResult

Adds a new relayer to the set

pub fn unregister_relayer(relayer: T::AccountId) -> DispatchResult

Removes a relayer from the set

pub fn transfer_fungible(
dest_id: ChainId,
resource_id: ResourceId,
to: Vec<u8>,
amount: U256
) -> DispatchResult

Initiates a transfer of a fungible asset out of the chain. This should be called by another pallet.

pub fn transfer_nonfungible(
dest_id: ChainId,
resource_id: ResourceId,
token_id: Vec<u8>,
to: Vec<u8>,
metadata: Vec<u8>
) -> DispatchResult

Initiates a transfer of a nonfungible asset out of the chain. This should be called by another pallet.

pub fn transfer_generic(
dest_id: ChainId,
resource_id: ResourceId,
metadata: Vec<u8>
) -> DispatchResult

Initiates a transfer of generic data out of the chain. This should be called by another pallet.

impl<T: Config> Pallet<T>

pub fn chains<KArg>(k: KArg) -> Option<DepositNonce> where
KArg: EncodeLike<ChainId>,

All whitelisted chains and their respective transaction counts

impl<T: Config> Pallet<T>

pub fn relayer_threshold() -> u32

Number of votes required for a proposal to execute

impl<T: Config> Pallet<T>

pub fn relayers<KArg>(k: KArg) -> bool where
KArg: EncodeLike<T::AccountId>,

Tracks current relayer set

impl<T: Config> Pallet<T>

pub fn relayer_count() -> u32

Number of relayers in set

impl<T: Config> Pallet<T>

pub fn votes<KArg1, KArg2>(
k1: KArg1,
k2: KArg2
) -> Option<ProposalVotes<T::AccountId, T::BlockNumber>> where
KArg1: EncodeLike<ChainId>,
KArg2: EncodeLike<(DepositNonce, T::Proposal)>,

All known proposals. The key is the hash of the call and the deposit ID, to ensure it’s unique.

impl<T: Config> Pallet<T>

pub fn resources<KArg>(k: KArg) -> Option<Vec<u8>> where
KArg: EncodeLike<ResourceId>,

Utilized by the bridge software to map resource IDs to actual methods