Trait bholdus_chainbridge::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type AdminOrigin: EnsureOrigin<Self::Origin>;
    type Proposal: Parameter + Dispatchable<Origin = Self::Origin> + EncodeLike + GetDispatchInfo;
    type ChainIdentity: Get<ChainId>;
    type ProposalLifetime: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

type AdminOrigin: EnsureOrigin<Self::Origin>

Origin used to administer the pallet

type Proposal: Parameter + Dispatchable<Origin = Self::Origin> + EncodeLike + GetDispatchInfo

Proposed dispatchable call

type ChainIdentity: Get<ChainId>

The identifier for this chain. This must be unique and must not collide with existing IDs within a set of bridged chains.

type ProposalLifetime: Get<Self::BlockNumber>

Loading content...